% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward-compat.R, R/cal_AUC_ml_res.R
\name{cal_auc_ml_res}
\alias{cal_auc_ml_res}
\alias{cal_AUC_ml_res}
\title{Calculate AUC scores of Machine Learning Models in all data}
\usage{
cal_auc_ml_res(...)

cal_AUC_ml_res(
  res.by.ML.Dev.Prog.Sig = NULL,
  train_data,
  inputmatrix.list,
  mode = NULL,
  AUC_time = 1,
  single_ml = NULL,
  double_ml1 = NULL,
  double_ml2 = NULL,
  alpha_for_Enet = NULL,
  direction_for_stepcox = NULL,
  auc_cal_method = "KM"
)
}
\arguments{
\item{...}{Arguments passed to `cal_AUC_ml_res()`.}

\item{res.by.ML.Dev.Prog.Sig}{the results of function ML.Dev.Prog.Sig}

\item{train_data}{the training data using in ML.Dev.Prog.Sig}

\item{inputmatrix.list}{A list containing data frames with columns ID, OS.time, OS, and model genes; expression values should be log2(x+1) scaled.}

\item{mode}{Choose MF models: all, single, double}

\item{AUC_time}{Time horizon in years, for example 1, 2, or 3. Use a horizon shorter than the minimum maximum survival time across cohorts.}

\item{single_ml}{If the mode is set to "single", you must fill in the following models: c("RSF", "Enet", "StepCox","CoxBoost","plsRcox","superpc","GBM","survivalsvm","Ridge","Lasso").}

\item{double_ml1}{If the mode is set to "double", you need to fill in the modeling methods here: c('RSF', "StepCox","CoxBoost","Lasso").}

\item{double_ml2}{If the mode is set to "double", you need to fill in the modeling methods here: c("RSF", "Enet", "StepCox","CoxBoost","plsRcox","superpc","GBM","survivalsvm","Ridge","Lasso")}

\item{alpha_for_Enet}{0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9}

\item{direction_for_stepcox}{c("both", "backward", "forward")}

\item{auc_cal_method}{KM, NNE}
}
\value{
AUC Calculation Result
}
\description{
Calculate AUC scores of Machine Learning Models in all data
}
\examples{
\dontrun{
all.auc.1y <- cal_AUC_ml_res(
  res.by.ML.Dev.Prog.Sig = res,
  train_data = list_train_vali_Data[["TCGA"]],
  inputmatrix.list = list_train_vali_Data, mode = "all",
  AUC_time = 1,
  auc_cal_method = "KM"
)

}
}
