% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Correlation between two genes}
\usage{
cor_plot(obj, dataset, color = NULL, feature1, feature2, method)
}
\arguments{
\item{obj}{A data frame of gene expression}

\item{dataset}{Identical name of dataset}

\item{color}{If NULL, color values are set to the default colors. Otherwise, you can specify one color value for line}

\item{feature1}{A specific gene}

\item{feature2}{Another specific gene}

\item{method}{Methods for correlation analysis, it can be "pearson" or "spearman"}
}
\value{
a ggplot2 object
}
\description{
Creates a scatter plot between two genes
}
\examples{
\dontrun{
# See package documentation for full workflows.
}
}
