% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exc.sig}
\alias{exc.sig}
\title{Exclusion signature gene sets}
\format{
See source data object.
}
\usage{
exc.sig
}
\description{
Gene sets used by immunotherapy response helpers.
}
\keyword{datasets}
