% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gp_IMPRES}
\alias{gp_IMPRES}
\title{Gene pairs for IMPRES immunotherapy response prediction}
\format{
See source data object.
}
\usage{
gp_IMPRES
}
\description{
Gene-pair definitions used by IMPRES-related scoring helpers.
}
\keyword{datasets}
