% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ImmmunCells.Sig}
\alias{ImmmunCells.Sig}
\title{Immune cell signature gene sets}
\format{
See source data object.
}
\usage{
ImmmunCells.Sig
}
\description{
Immune cell signatures used by immunotherapy response helpers.
}
\keyword{datasets}
