% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPRES_gmt}
\alias{IPRES_gmt}
\title{IPRES gene set (GMT format)}
\format{
See source data object.
}
\usage{
IPRES_gmt
}
\description{
IPRES signature represented in GMT-style data.
}
\keyword{datasets}
