% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NLRP3.Sig.gmt}
\alias{NLRP3.Sig.gmt}
\title{NLRP3 inflammasome signature (GMT format)}
\format{
See source data object.
}
\usage{
NLRP3.Sig.gmt
}
\description{
NLRP3 inflammasome signature represented in GMT-style data.
}
\keyword{datasets}
