% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pre.prog.sig}
\alias{pre.prog.sig}
\title{Pre-defined prognostic signatures}
\format{
See source data object.
}
\usage{
pre.prog.sig
}
\description{
Curated prognostic signatures used by previous-signature helpers.
}
\keyword{datasets}
