% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_vis_category.R
\name{roc_vis_category}
\alias{roc_vis_category}
\title{Receiver operating characteristic curves when predicting response}
\usage{
roc_vis_category(
  object,
  model_name,
  dataset,
  order = NULL,
  dataset_col = NULL,
  anno_position = NULL
)
}
\arguments{
\item{object}{Output of function ML.Dev.Pred.Category.Sig}

\item{model_name}{Model name in object to plot}

\item{dataset}{Identical name of dataset in object to plot}

\item{order}{If NULL, order is set to the default order. Otherwise, you can specify the order of datasets to plot}

\item{dataset_col}{If NULL, color values are set to the default colors. Otherwise, you can specify consistent number of color value for cohorts}

\item{anno_position}{The position of AUC value}
}
\value{
a ggplot2 object
}
\description{
Creates a plot of ROC for specific method among different datasets when predicting response
}
\examples{
\dontrun{
# See package documentation for full workflows.
}
}
