% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TcellExc.Sig}
\alias{TcellExc.Sig}
\title{T cell exclusion signature gene sets}
\format{
See source data object.
}
\usage{
TcellExc.Sig
}
\description{
T cell exclusion signatures used by immunotherapy response helpers.
}
\keyword{datasets}
