% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRS.Sig.gmt}
\alias{TRS.Sig.gmt}
\title{Tumor resistance signature (GMT format)}
\format{
See source data object.
}
\usage{
TRS.Sig.gmt
}
\description{
Tumor resistance signature represented in GMT-style data.
}
\keyword{datasets}
