% Generated by roxygen2: do not edit by hand
\name{gp_IMPRES}
\alias{gp_IMPRES}
\docType{data}
\title{Gene pairs for IMPRES immunotherapy response prediction.}
\description{
Gene pairs for IMPRES immunotherapy response prediction.
}
\usage{data(gp_IMPRES)}
\format{See source documentation.}
\keyword{datasets}
